add health_check endpoint, add test
