+++ Cargo.toml
@@ -5,6 +5,13 @@ authors = ["Takehiro Ohtani <isit7a@gmail.com>"]
 edition = "2018"
 
 # See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
+[lib]
+path = "src/lib.rs"
+name = "lib"
+
+[[bin]]
+path = "src/bin/main.rs"
+name = "main"
 
 [dependencies]
 actix-web = "3"
@@ -19,3 +26,6 @@ num_cpus = "1.13.0"
 once_cell = "1.7.2"
 serde = "1.0.125"
 serde_json = "1.0.64"
+
+[dev-dependencies]
+actix-rt = "1.1.1"
