+++ src/schema.rs
@@ -19,4 +19,7 @@ table! {
 
 joinable!(users -> companies (company_id));
 
-allow_tables_to_appear_in_same_query!(companies, users,);
+allow_tables_to_appear_in_same_query!(
+    companies,
+    users,
+);
